% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin_covariates}
\alias{admin_covariates}
\title{Covariates at admin unit 1 level.}
\format{a data frame with 3305 rows and 37 columns}
\usage{
admin_covariates
}
\description{
A dataset containing global environmental and demographic variables at
administrative unit 1 resolution.
}
\keyword{datasets}
