% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_structure}
\alias{age_structure}
\title{Countries' age structure}
\format{a data frame with 200 rows and 28 columns}
\usage{
age_structure
}
\description{
A dataset containing the proportion of individuals in 20 5-year age groups for 200 countries.
}
\keyword{datasets}
