% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_sqr_covariates}
\alias{all_sqr_covariates}
\title{Force of infection estimates and covariates at 1/6 degree.}
\format{a data frame with 425138 rows and 43 columns}
\usage{
all_sqr_covariates
}
\description{
A dataset containing FOI estimates and environmental and demographic covariates at
1/6 degree resolution.
}
\keyword{datasets}
