% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Brazil_contour}
\alias{Brazil_contour}
\title{Brazil national administrative boundary.}
\format{a sf object with 1 row}
\usage{
Brazil_contour
}
\description{
A shapefile of the GADM national administrative boundary of Brazil.
}
\keyword{datasets}
