% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_calculating_burden.R
\name{create_lookup_tables}
\alias{create_lookup_tables}
\title{Create look up tables for burden and R0 calculation}
\usage{
create_lookup_tables(i, age_struct, age_band_tags, age_band_L_bounds,
  age_band_U_bounds, parms)
}
\arguments{
\item{age_struct}{dataframe of the proportion of individuals in different 5-yr age groups,
for each country.}

\item{age_band_tags}{character string of the names of the columns of the `age_structure` dataframe
containing the actual age structure data.}

\item{age_band_L_bounds}{integers of the lower limits of the age groups.}

\item{age_band_U_bounds}{integers pf the upper limits of the age groups.}

\item{parms}{list of user-defined parameters.}
}
\description{
The function generates look up tables for calculating total number of annual dengue infections,
 mild febrile cases, cases requiring hospitalization and R0. R0 is estimated for
 two different assumptions of the infectiousness of the four dengue infections.
}
