% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_list.R
\name{create_parameter_list}
\alias{create_parameter_list}
\title{Create a list of user-defined parameters}
\usage{
create_parameter_list(no_predictors = 16, dependent_variable = "FOI",
  resample_grid_size = 20, grid_size = 5, no_samples = 1,
  vec_phis_R0_1 = c(1, 1, 0, 0), vec_phis_R0_2 = c(1, 1, 1, 1),
  prop_sympt = c(0.45, 0.85, 0.15, 0.15), prop_hosp = c(0.04, 0.1,
  0.04, 0.04), FOI_grid = seq(0.002, 0.2, 0.002), sf_vals = seq(1, 0.1,
  -0.1), sat_functions_shapes = c(0, 5, 1600000), no_trees = 500,
  min_node_size = 20, all_wgt = 1, pseudoAbs_value = c(FOI = -0.02,
  R0_1 = 0.5, R0_2 = 0.5, R0_3 = 0.5, Z = -0.02), foi_offset = 0.03,
  EM_iter = 10, base_info = c("cell", "latitude", "longitude",
  "population", "ID_0", "ID_1", "ID_2"), extra_params = NULL)
}
\arguments{
\item{no_predictors}{number of selected covariates used for model fitting and making predictions.}

\item{dependent_variable}{character string of the fitted response variable (FOI).}

\item{resample_grid_size}{resolution of the foi predictions (in km). Default = 20 (1/6 degree).}

\item{grid_size}{size of the grid used for block bootstrapping.}

\item{no_samples}{number of bootstrap samples.}

\item{vec_phis_R0_1}{numeric vector of length = 4 of the relative infectiousness of primary,
secondary, tertiary and quaternary dengue infections, when assuming only primary and
secondary infections are infectious.}

\item{vec_phis_R0_2}{numeric vector of length = 4 of the relative infectiousness of primary,
secondary, tertiary and quaternary dengue infections, when assuming all four infections
are infectious.}

\item{prop_sympt}{numeric vector of length = 4 of the proportions of primary, secondary,
tertiary and quaternary infections which are symptomatic.}

\item{prop_hosp}{numeric vector of length = 4 of the proportions of primary, secondary,
tertiary and quaternary infections requiring hospitalization.}

\item{FOI_grid}{numeric vector of force of infection values used for mapping FOI to
number of infections, cases hopsitalizations and R0.}

\item{sf_vals}{scaling factor used to model the effect of transmission-reducing interventions.}

\item{sat_functions_shapes}{parameters of the saturating function used for setting pseudo absences
case weights.}

\item{no_trees}{number of trees of the random forest model passed to \code{\link{ranger}}.}

\item{min_node_size}{minimal node size of the random forest model passed to \code{\link{ranger}}.}

\item{all_wgt}{numeric value of case weights for all data points.}

\item{pseudoAbs_value}{numeric value of pseudo absences for different response variables.}

\item{foi_offset}{numeric value to offset model FOI predictions during fitting with
Expectation Maximization.}

\item{base_info}{character string of key ID variable in the 1/6 degree resolution global
covariate dataset.}

\item{extra_params}{list of additional parameters. Default = NULL.}
}
\description{
The function creates a list of user-defined parameters.
}
