% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endemic_ID_0_ID_1}
\alias{endemic_ID_0_ID_1}
\title{Dengue endemic countries.}
\format{a data frame with 1952 rows and 3 columns}
\usage{
endemic_ID_0_ID_1
}
\description{
A dataset containing GADM identifier fields for country and administrative unit 1
where dengue is an endemic disease.
}
\keyword{datasets}
