% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_max_algorithm.R
\name{exp_max_algorithm}
\alias{exp_max_algorithm}
\title{\code{exp_max_algorithm} fit a random forest model to one single
  bootstrap sample of the data using the EM algorithm}
\usage{
exp_max_algorithm(parms, adm_dataset, pxl_dataset, covariates_names,
  map_col = NULL, RF_obj_path = NULL, RF_obj_name = NULL,
  diagn_tab_path = NULL, diagn_tab_name = NULL, map_path = NULL,
  sct_plt_path = NULL, train_dts_path = NULL, train_dts_name = NULL,
  adm_covariates = NULL)
}
\arguments{
\item{parms}{list of user-defined parameters.}

\item{adm_dataset}{dataframe of the bootstrapped dataset of foi estimates and
covariates at admin unit 1 resolution (pre-processed).}

\item{pxl_dataset}{dataframe of the bootstrapped dataset of foi estimates and
covariates at 1/6 degree resolution.}

\item{covariates_names}{character vector of covariates names.}

\item{adm_covariates}{dataframe of global covariates at admin unit 1 resolution.}
}
\value{
the random forest model object returned by \code{ranger}
}
\description{
\code{exp_max_algorithm} fit a random forest model to one single
  bootstrap sample of the data using the EM algorithm
}
