% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ranger_RF_and_make_predictions.R
\name{fit_ranger_RF}
\alias{fit_ranger_RF}
\title{\code{fit_ranger_RF} fit a random forest model using \code{ranger}.}
\usage{
fit_ranger_RF(parms, dependent_variable, covariates_names,
  training_dataset, my_weights)
}
\arguments{
\item{parms}{list of user-defined parameters.}

\item{dependent_variable}{character string of the fitted response variable (FOI).}

\item{covariates_names}{character vector of covariates names.}

\item{training_dataset}{a dataframe of the dataset used for model training.}

\item{my_weights}{character string of the name of the column of \code{training dataset}
with case weights.}

\item{sqr_dataset_2}{dataframe of foi estimates and predictions at 1/6 degree resolution.}
}
\value{
the random forest model object returned by \code{ranger}
}
\description{
\code{fit_ranger_RF} fit a random forest model using \code{ranger}.
}
