% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_up.R
\name{fitted_sero_cell_to_adm}
\alias{fitted_sero_cell_to_adm}
\title{Modify predictions of serology data points at 1/6 degree resolution}
\usage{
fitted_sero_cell_to_adm(join_all, sqr_dataset_2, fld_1, fld_2)
}
\arguments{
\item{join_all}{dataframe of foi estimates and predictions at admin unit 1 resolution.}

\item{sqr_dataset_2}{dataframe of foi estimates and predictions at 1/6 degree resolution.}

\item{fld_1}{character strings of variables (columns names)}

\item{fld_2}{character strings of variables (columns names)}
}
\value{
a joined dataframe
}
\description{
The function replaces the mean of all 1/6 degree predictions within adim unit 1 level
serology data with the value of the prediction in 1/6 degree resolution cell
where the original serology data point belongs. This allows to fit the
serology data points as point data.
}
