% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foi_data}
\alias{foi_data}
\title{Force of infection estimates and covariates at admin unit 1.}
\format{a data frame with 693 rows and 49 columns}
\usage{
foi_data
}
\description{
A dataset containing FOI estimates and environmental and demographic covariates at
administrative unit 1 resolution.
}
\keyword{datasets}
