% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_routine_bootstrap.R
\name{full_routine_bootstrap}
\alias{full_routine_bootstrap}
\title{Pre-process and fit a random forest model to a sample of FOI data}
\usage{
full_routine_bootstrap(parms, original_foi_data, adm_covariates,
  all_squares, covariates_names, boot_sample)
}
\arguments{
\item{parms}{list of user-defined parameters.}

\item{original_foi_data}{dataframe of the original foi estimates dataset at
admin unit 1 resolution.}

\item{adm_covariates}{dataframe of global covariates at admin unit 1 resolution. Default = NULL.}

\item{all_squares}{dataframe of global covariates at 1/6 degree resolution.}

\item{covariates_names}{character vector of covariates names.}

\item{boot_sample}{dataframe of the bootstrapped dataset of foi estimates and
covariates at admin unit 1 resolution (not pre-processed).}
}
\value{
the random forest model object returned by \code{\link{ranger}}.
}
\description{
The function preprocess one single bootstrap sample of the original FOI data
and fits a random forest model to it using the EM algorithm.
}
