% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_scatter_plot.R
\name{generic_scatter_plot}
\alias{generic_scatter_plot}
\title{Plot observations vs predictions}
\usage{
generic_scatter_plot(df, x, y, out_name, out_pt)
}
\arguments{
\item{df}{dataframe admin unit 1 level FOI observations and predictions.}

\item{x}{character of the column name containing the observations.}

\item{y}{character of the column name containing the predictions.#}
}
\description{
The function generates a plot of admin unit 1 level FOI observations against
predictions with the R2 shown.
}
