% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pseudo_abs_weights.R
\name{get_sat_area_wgts}
\alias{get_sat_area_wgts}
\title{Calculate case weights for admin unit 1 level pseudo absences}
\usage{
get_sat_area_wgts(foi_data, parms)
}
\arguments{
\item{foi_data}{dataframe of admin unit level FOI data.}

\item{parms}{list of user-defined parameters.}
}
\description{
The function calculates case weights for admin unit 1 level pseudo absence points
assuming that the case weight value is a saturating function of the area (km) of
the admin unit.
}
