% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{loop}
\alias{loop}
\title{Parallel computation wrapper}
\usage{
loop(..., parallel)
}
\arguments{
\item{...}{aaditional arguments passed on to lapply or parLapply.}

\item{parallel}{logical. Is parLapply used?}
}
\description{
The function is a wrapper to lapply and parLapply.
}
