% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster_map.R
\name{map_preprocess}
\alias{map_preprocess}
\title{Preprocess data for ggplot map}
\usage{
map_preprocess(pred2, var_to_plot, parms)
}
\arguments{
\item{pred2}{dataframe of data.}

\item{var_to_plot}{character of the name of the variable to plot.}

\item{parms}{list of user-defined parameters.}
}
\description{
The function assignes equally spaced latitude and longitude coordinates to
each 1/6 degree pixel and shapes the output for making a map using \code{\link{ggplot}}.
}
