% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_to_vaccine_impact_calculation.R
\name{pre_process_vaccine_lookup_table}
\alias{pre_process_vaccine_lookup_table}
\title{Pre-process the vaccine impact look up table}
\usage{
pre_process_vaccine_lookup_table(look_up_table, R0_preds)
}
\arguments{
\item{look_up_table}{original look up table of vaccine impact}

\item{R0_preds}{numeric vector of R0 predictions}
}
\description{
The function edits the original look up table of the impact of the Sanofi Pasteur
dengue vaccine for different R0 values, as output by the Ferguson et al 2016
(Science) model. Specifically it adds to it the R0 -> vaccine impact for R0 = 0
and for R0 = max R0 predicted by the random forest model.
It also removes some columns of original vaccine impact lookup table.
}
