% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictor_rank}
\alias{predictor_rank}
\title{Selected predictors.}
\format{a data frame with 28 rows and 3 columns}
\usage{
predictor_rank
}
\description{
A dataset containing the top explanatory variables as output from variable sslection.
}
\keyword{datasets}
