% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_adm_data}
\alias{preprocess_adm_data}
\title{Pre-process the original admin unit level FOI dataset}
\usage{
preprocess_adm_data(parms, foi_data)
}
\arguments{
\item{parms}{list of user-defined parameters.}

\item{foi_data}{dataframe of admin unit level FOI data.}
}
\description{
The function sets the case weigths and applies an offset to the FOI response variable
(offset is needed during the EM fitting) in the original admin unit level FOI dataset.
}
