% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_pxl_data}
\alias{preprocess_pxl_data}
\title{Pre-process the 1/6 degree resolution FOI dataset}
\usage{
preprocess_pxl_data(parms, foi_data, pxl_data)
}
\arguments{
\item{parms}{list of user-defined parameters.}

\item{foi_data}{dataframe of admin unit level FOI data.}

\item{pxl_data}{dataframe of 1/6 degree resolution FOI data.}
}
\description{
The function joins the original admin unit level FOI estimates to the 1/6 degree resolution
FOI dataset and also sets the case weigths for each 1/6 degree resolution data point.
}
