% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster_map.R
\name{quick_raster_map}
\alias{quick_raster_map}
\title{Plot and save a raster map}
\usage{
quick_raster_map(pred_df, statistic, my_col, out_pt, out_name,
  z_range = NULL, shp = NULL, key_ttle = NULL)
}
\arguments{
\item{pred_df}{dataframe of data.}

\item{statistic}{character of the name of the variable to plot.}

\item{my_col}{colours to use for plotting.}

\item{out_pt}{output directory.}

\item{out_name}{output file name.}

\item{z_range}{range of values to show on plot. Default = NULL.}

\item{shp}{shapefile object. Default = NULL.}

\item{key_ttl}{character string of legend title. Default = NULL.}
}
\description{
The function makes and saves locally a raster map using \code{\link{image.plot}}.
}
