% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R0_to_prop_cases_averted_lookup_1}
\alias{R0_to_prop_cases_averted_lookup_1}
\title{Look up table of vaccine impact on mild cases (R0 assumption 1).}
\format{a data frame with 17 rows and 19 columns}
\usage{
R0_to_prop_cases_averted_lookup_1
}
\description{
A dataset containing the fraction of mild febbrile cases averted following application of the
Sanofi Pasteur dengue vaccine, when assuming only primary and secondary
infections are infectious, for different R0 values.
}
\keyword{datasets}
