% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R0_to_prop_hosp_averted_lookup_2}
\alias{R0_to_prop_hosp_averted_lookup_2}
\title{Look up table of vaccine impact on hospitalized cases (R0 assumption 2).}
\format{a data frame with 17 rows and 19 columns}
\usage{
R0_to_prop_hosp_averted_lookup_2
}
\description{
A dataset containing the fraction of hospitalized cases averted following application of the
Sanofi Pasteur dengue vaccine, when assuming all 4 infections are infectious,
for different R0 values.
}
\keyword{datasets}
