% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pseudo_abs_weights.R
\name{set_wgts_to_sero_cells}
\alias{set_wgts_to_sero_cells}
\title{Calculate case weights for serology 1/6 degree resolution cells}
\usage{
set_wgts_to_sero_cells(foi_data, pxl_data, parms)
}
\arguments{
\item{foi_data}{dataframe of admin unit level FOI data.}

\item{pxl_data}{dataframe of 1/6 degree resolution FOI data.}

\item{parms}{list of user-defined parameters.}
}
\description{
The function sets to 1 the case weights of the 1/6 degree resolution cells where
the original serology data points were located. It sets to 0 the case weights of
all the other 1/6 degree resolution cells in the data points admin unit. This
allows to fit the serology data points as point data.
}
