% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_to_replicate_R0_and_burden.R
\name{wrapper_to_replicate_R0_and_burden}
\alias{wrapper_to_replicate_R0_and_burden}
\title{Look up the impact of a transmission-reducing intervention}
\usage{
wrapper_to_replicate_R0_and_burden(foi_data, age_struct, scaling_factor,
  FOI_to_Inf_list, FOI_to_C_list, FOI_to_HC_list, FOI_to_R0_1_list,
  FOI_to_R0_2_list, age_band_lower_bounds, age_band_upper_bounds,
  age_band_tags, parms)
}
\arguments{
\item{foi_data}{dataframe of processed admin unit level foi data.}

\item{age_struct}{dataframe of the proportion of individuals in different 5-yr age groups,
for each country.}

\item{scaling_factor}{numeric scaling factor value (0-1)}

\item{FOI_to_Inf_list}{list of FOI -> number of infections look up tables for different countries}

\item{FOI_to_C_list}{list of FOI -> number of cases look up tables for different countries}

\item{FOI_to_HC_list}{list of FOI -> number of hospitalized cases look up tables for different countries}

\item{FOI_to_R0_1_list}{list of FOI -> R0 (assumption 1) look up tables for different countries}

\item{FOI_to_R0_2_list}{list of FOI -> R0 (assumption 2) look up tables for different countries}

\item{age_band_tags}{character string of the names of the columns of the `age_structure` dataframe
containing the actual age structure data.}

\item{parms}{list of user-defined parameters.}
}
\description{
The function look up the two R0 values, the number of infections, mild febrile cases and
the number of hospitalized cases corresponding to a set of FOI predictions,
following the application of a scaling factor representing the assumed
reduction in transmission.
}
