% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_to_vaccine_impact_calculation.R
\name{wrapper_to_replicate_vaccine_impact}
\alias{wrapper_to_replicate_vaccine_impact}
\title{Look up the impact of the Sanofi Pasteur dengue vaccine}
\usage{
wrapper_to_replicate_vaccine_impact(R0_preds, look_up_table, screen_age,
  parms)
}
\arguments{
\item{R0_preds}{numeric vector of R0 predictions}

\item{look_up_table}{original look up table of vaccine impact}

\item{screen_age}{integer value of assumed screening age of vaccination}

\item{parms}{list of user-defined parameters.}
}
\description{
The function looks up the vaccine impact (i.e. estimated proportion of infections,
cases or hospitalizations averted) corresponding to a set of R0 predictions.
}
