% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wgt_corr.R
\name{calculate_wgt_cor}
\alias{calculate_wgt_cor}
\title{Calculate weighted Pearson correlation coefficient}
\usage{
calculate_wgt_cor(d.sub, x, y)
}
\arguments{
\item{x}{character string of the name of variable A.}

\item{y}{character string of the name of variable B.}

\item{d.subs}{a dataframe with data.}
}
\value{
the numeric value of the correlation coefficient.
}
\description{
The function calculates a weighted Pearson correlation coefficient between
variable A and B using \code{\link{wtd.cor}}.
}
