% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_creating_bootstrap_samples.R
\name{grid_and_boot}
\alias{grid_and_boot}
\title{\code{grid_and_boot} take a block bootstrapped sample of the original foi dataset.}
\usage{
grid_and_boot(data_df, parms)
}
\arguments{
\item{data_df}{dataframe of foi predictions dataset at admin unit 1 resolution.}

\item{parms}{list of user-defined parameters.}
}
\description{
\code{grid_and_boot} take a block bootstrapped sample of the original foi dataset.
}
