% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{loop}
\alias{loop}
\title{Wrapper for parallel computation}
\usage{
loop(..., parallel)
}
\arguments{
\item{...}{aerguement passed on to \code{lapply} or \code{parallel::parLapply}.}

\item{parallel}{logicla indicating whether to run things in parallel.}
}
\description{
The function allows to do parallel computation using the \code{\link{parallel}} package
}
