% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster_map.R
\name{quick_raster_map}
\alias{quick_raster_map}
\title{Plot and save a raster map}
\usage{
quick_raster_map(pred_df, my_col, ttl = NULL, parms)
}
\arguments{
\item{pred_df}{dataframe of data.}

\item{my_col}{colours to use for plotting.}

\item{ttl}{character string of legend title. Default = NULL.}

\item{parms}{list of user-defined parameters.}
}
\description{
The function makes and saves locally a raster map using \code{\link{ggplot}}.
}
